sagemaker/requirements.txt
boto3
wandb
python-dotenv
pyyaml
requests
ultralytics
torch
pillow
numpy
opencv-python
fastapi
uvicorn 